server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-rest-knox==4.2.0
pyotp==2.8.0
cryptography==41.0.5
django-cors-headers
qrcode
gunicorn==21.2.0